requirements.txt
accelerate==1.9.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.9.0
asttokens
autopep8
av==15.0.0
backoff==2.2.1
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
comm
contourpy==1.3.3
cycler==0.12.1
debugpy
decorator
einops==0.8.1
exceptiongroup
executing
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.18.0
# flash_attn: install separately, see https://github.com/Dao-AILab/flash-attention
Flask==3.1.2
fonttools==4.60.1
fsspec==2025.7.0
GPUtil==1.4.0
gradio==5.4.0
gradio_client==1.4.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.10
importlib_metadata
ipykernel
ipython
ipython_pygments_lexers
itsdangerous==2.2.0
jedi
Jinja2==3.1.6
jupyter_client
jupyter_core
kiwisolver==1.4.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline
mdurl==0.1.2
mpmath==1.3.0
nest_asyncio
networkx==3.5
numpy==2.3.1
orjson==3.11.0
packaging
pandas==2.3.1
parso
peft==0.18.1
pexpect
pickleshare
pillow==11.3.0
platformdirs
prompt_toolkit
protobuf==6.31.1
psutil
ptyprocess
pure_eval
pycodestyle
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments
pyparsing==3.2.5
python-dateutil
python-multipart==0.0.12
pytz==2025.2
PyYAML==6.0.2
pyzmq
qwen-vl-utils==0.0.10
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruff==0.12.4
safehttpx==0.1.6
safetensors==0.5.3
scipy==1.17.0
semantic-version==2.10.0
sentencepiece==0.2.1
setuptools==78.1.1
shellingham==1.5.4
six
sniffio==1.3.1
stack_data
starlette==0.47.2
sympy==1.14.0
tokenizers==0.22.1
tomlkit==0.12.0
# torch, torchvision: install separately based on your CUDA version, see https://pytorch.org/get-started/locally/
tornado
tqdm==4.67.1
traitlets
transformers==4.57.3
transformers-stream-generator==0.0.4
tueplots==0.2.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritools==5.0.0
urlextract==1.9.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth
websockets==12.0
Werkzeug==3.1.5
wheel==0.45.1
zipp
